+++ Documentation/Lambda Calculus.playground/contents.xcplayground
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='3.0' sdk='macosx'>
+<playground version='3.0' sdk='macosx' auto-termination-delay='5'>
     <sections>
         <code source-file-name='section-1.swift'/>
     </sections>
