--- Documentation/Madness.playground/timeline.xctimeline
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Timeline
-   version = "3.0">
-   <TimelineItems>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "#CharacterRangeLen=0&amp;CharacterRangeLoc=0&amp;EndingColumnNumber=10&amp;EndingLineNumber=37&amp;StartingColumnNumber=1&amp;StartingLineNumber=37&amp;Timestamp=440094530.375359">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "#CharacterRangeLen=66&amp;CharacterRangeLoc=1333&amp;EndingColumnNumber=68&amp;EndingLineNumber=50&amp;StartingColumnNumber=2&amp;StartingLineNumber=50&amp;Timestamp=440094536.912156">
-      </LoggerValueHistoryTimelineItem>
-   </TimelineItems>
-</Timeline>
